bora

+++ app/Http/Controllers/Admin/PostController.php
@@ -4,13 +4,19 @@ namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
 use App\Http\Requests\PostRequest;
+use App\Models\Category;
+use App\Models\CategoryPost;
 use App\Models\Post;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class PostController extends Controller
 {
         
     public function list(Request $request){
+        Gate::authorize('viewAny', Post::class);
+        
         $pagination = Post::orderBy("subject");
 
         if (isset($request->busca) && $request->busca != "") {
@@ -33,22 +39,35 @@ class PostController extends Controller
     }
 
     public function create(){
-        return view("admin.posts.form", ["data"=>new Post()] );
+        Gate::authorize('create', Post::class);
+        $categoriesList = Category::all();
+        return view("admin.posts.form", ["data"=>new Post(),
+                                        "categoriesList"=>$categoriesList] );
     }
 
     public function store(PostRequest $request){
+        Gate::authorize('create', Post::class);
         $validated = $request->validated();
 
         $path = $request->file('image')->store('posts',"public");
 
         $data = $request->all();
         $data["image"] = $path;
+        $data["user_id"] = Auth::user()->id;
 
         $post = Post::create($data);
+
+        
+        #vinculação com categoria
+        $cat = Category::find($request["category_id"]);
+        CategoryPost::updateOrCreate(["post_id"=>$post->id,"category_id"=>$cat->id]);
+    
+
         return redirect(route("post.edit", $post))->with("success",__("Data saved!"));
     }
 
     public function destroy(Post $post){
+        Gate::authorize('delete', $post);
         $post->delete();
         return redirect(route("post.list"))->with("success",__("Data deleted!"));
     }
@@ -56,11 +75,22 @@ class PostController extends Controller
 
     #abre o formulario de edição
     public function edit(Post $post){
-        return view("admin.posts.form",["data"=>$post]);
+        Gate::authorize('view', $post);
+        $categoriesList = Category::all();
+
+        $categories = Category::select("categories.*", "category_posts.id as category_posts_id")
+                        ->join("category_posts","category_posts.category_id","=","categories.id")
+                        ->where("post_id",$post->id)->paginate(2);
+
+
+        return view("admin.posts.form",["data"=>$post,
+                                        "categoriesList"=>$categoriesList,
+                                        "categories"=>$categories]);
     }
 
     #salva as edições
     public function update(Post $post, PostRequest $request) {
+        Gate::authorize('update', $post);
         $validated = $request->validated();
 
         $data = $request->all();
@@ -71,6 +101,15 @@ class PostController extends Controller
         }
 
         $post->update($data);
+
+
+        #vinculação com categoria
+        if ($request["category_id"]){
+            $cat = Category::find($request["category_id"]);
+            CategoryPost::updateOrCreate(["post_id"=>$post->id,"category_id"=>$cat->id]);
+        }
+
+
         return redirect()->back()->with("success",__("Data updated!"));
     }
